requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
astrapy==0.6.2
asyncer==0.0.2
attrs==23.1.0
backoff==2.2.1
bidict==0.22.1
blinker==1.7.0
cachetools==5.3.2
cassandra-driver==3.29.0
cassio==0.1.3
certifi==2023.11.17
chainlit==0.7.700
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
Deprecated==1.2.14
distro==1.9.0
fastapi==0.100.1
fastapi-socketio==0.0.10
filetype==1.2.0
frozenlist==1.4.1
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.40
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.6
importlib-metadata==6.11.0
inflection==0.5.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
langchain==0.0.352
langchain-community==0.0.6
langchain-core==0.1.3
langsmith==0.0.75
Lazify==0.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
monotonic==1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.26.2
openai==1.6.1
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-anthropic==0.7.0
opentelemetry-instrumentation-bedrock==0.7.0
opentelemetry-instrumentation-chromadb==0.7.0
opentelemetry-instrumentation-cohere==0.7.0
opentelemetry-instrumentation-dbapi==0.42b0
opentelemetry-instrumentation-langchain==0.7.0
opentelemetry-instrumentation-llamaindex==0.7.0
opentelemetry-instrumentation-openai==0.7.0
opentelemetry-instrumentation-pinecone==0.7.0
opentelemetry-instrumentation-pymysql==0.42b0
opentelemetry-instrumentation-replicate==0.7.0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-transformers==0.7.0
opentelemetry-instrumentation-urllib3==0.42b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-semantic-conventions-ai==0.0.12
opentelemetry-util-http==0.42b0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
posthog==3.1.0
protobuf==4.25.1
pyarrow==14.0.2
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.10.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
streamlit==1.29.0
syncer==2.0.3
tenacity==8.2.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traceloop-sdk==0.7.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uptrace==1.22.0
urllib3==2.1.0
uvicorn==0.23.2
validators==0.22.0
watchfiles==0.20.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
